airbyte-commons/src/test/java/io/airbyte/commons/io/LogTypeMapperTest.java
/*
 * MIT License
 *
 * Copyright (c) 2020 Airbyte
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

package io.airbyte.commons.io;

import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.verify;

import org.junit.jupiter.api.Test;
import org.slf4j.Logger;

class LogTypeMapperTest {

  @Test
  public void testDefaultLogger() {
    Logger logger = mock(Logger.class);
    String logString = "message not prefixed with a log level such as INFO";
    LogTypeMapper mapper = new LogTypeMapper(logger, l -> l::debug);

    mapper.accept(logString);
    verify(logger).debug(logString);
  }

  @Test
  public void testOverriddenLogger() {
    Logger logger = mock(Logger.class);
    String logString = "INFO message prefixed with a log level";
    LogTypeMapper mapper = new LogTypeMapper(logger, l -> l::debug);

    mapper.accept(logString);
    verify(logger).info(logString);
  }

}
